refactor(server): move snapshot service into package
