+++ packages/doc-server/src/share-web-routes.ts
@@ -13,7 +13,7 @@ import type { ShareRole } from './share-types.js';
 import { recordShareLinkOpen } from '../../../server/metrics.js';
 import { getCookie, shareTokenCookieName } from './cookies.js';
 import { handleShareMarkdown, shareMarkdownBodyParser } from '../../../server/routes.js';
-import { getSnapshotHtml, getSnapshotPublicUrl } from '../../../server/snapshot.js';
+import { getSnapshotHtml, getSnapshotPublicUrl } from './snapshot.js';
 import { stripProofSpanTags } from './proof-span-strip.js';
 import {
   getCanonicalReadableDocumentSync,
